Added JSON_QUERY and ISJSON methods
